refactor(ai-engine): resolve #1729 — group modpack agent cluster into modpack/ subpackage
Group the three related modpack conversion files (gameplay_comparison_agent,
modpack_conflict_detector, modpack_orchestrator) into a cohesive modpack/
subpackage under agents/, following the established split pattern (#1707-#1709).

Structure:
  ai-engine/agents/modpack/
  ├── __init__.py            # re-exports full public API
  ├── orchestrator.py        # from modpack_orchestrator.py
  ├── conflict_detector.py   # from modpack_conflict_detector.py
  └── gameplay_comparator.py # from gameplay_comparison_agent.py

- Reduced the three flat agents/*.py files to backward-compat re-export
  stubs so existing 'from agents.modpack_orchestrator import X' imports
  keep working (avoids the dangling-import pattern from #1819).
- Updated orchestrator's relative imports: conflict_detector is now a
  sibling (.), curseforge/modrinth/mod_dependency analyzers are parent (..).
- Updated all three test modules to import from the new canonical path.
- No modpack_types.py created: the three files own their dataclasses
  independently with no genuinely shared types; an empty shared module
  would be dead code.

Verification:
- 84 modpack/gameplay/conflict tests pass (3 test files).
- ruff check clean on new subpackage, stubs, and updated tests.
- No dangling old-style relative imports remain (grep verified).

Fixes #1729
